reverse-string: Moved the lints into the new lints.rs module
