+++ Cargo.toml
@@ -7,7 +7,6 @@ description = "Exercism's automated analyzer for the Rust track."
 [dependencies]
 failure = "0.1.5"
 serde_json = "1.0.39"
-lazy_static = "1.3.0"
 
 [dependencies.serde]
 version = "1.0.91"
