+++ src/analyzers/reverse_string/mod.rs
@@ -1,4 +1,5 @@
 pub mod comments;
+mod lints;
 #[cfg(test)]
 mod test;
 
@@ -10,223 +11,10 @@ use crate::{
     Result,
 };
 use comments::ReverseStringComment;
+use lints::REVERSE_STRING_LINTS;
 use syn::File;
 
 pub struct ReverseStringAnalyzer;
-
-type ReverseStringLint = Box<dyn Fn(&File, &mut AnalysisOutput) -> Result<()>>;
-
-/// Generates a vector of the linting functions. Each one of them accepts a reference
-/// to the solution AST and a mutable reference to the analysis output and
-/// modifies the output in accordance to the lint that they represent.
-fn generate_lints() -> Vec<ReverseStringLint> {
-    vec![
-        Box::new(|ast, output| {
-            let has_extern = ast.items.iter().any(|item| {
-                if let syn::Item::ExternCrate(_) = item {
-                    true
-                } else {
-                    false
-                }
-            });
-            if has_extern {
-                output
-                    .comments
-                    .push(ReverseStringComment::SuggestRemovingExternCrate.to_string());
-            }
-            Ok(())
-        }),
-        Box::new(|ast, output| {
-            let reverse_fn = if let Some(syn::Item::Fn(func)) = ast.items.iter().find(|item| {
-                if let syn::Item::Fn(item_fn) = item {
-                    item_fn.ident == "reverse"
-                } else {
-                    false
-                }
-            }) {
-                func
-            } else {
-                // We can use the unreachable!, since we have already check for
-                // the "reverse" function existence at the analyze function.
-                unreachable!();
-            };
-            // Check if the "reverse" function contains the "input.chars().rev().collect()"
-            // solution
-            let has_standard_solution = reverse_fn.block.stmts.iter().any(|stmt| {
-                if let syn::Stmt::Expr(syn::Expr::MethodCall(syn::ExprMethodCall {
-                    method,
-                    receiver,
-                    ..
-                })) = stmt
-                {
-                    if method != "collect" {
-                        false
-                    } else if let syn::Expr::MethodCall(syn::ExprMethodCall {
-                        method,
-                        receiver,
-                        ..
-                    }) = receiver.as_ref()
-                    {
-                        if method != "rev" {
-                            false
-                        } else if let syn::Expr::MethodCall(syn::ExprMethodCall {
-                            method,
-                            receiver,
-                            ..
-                        }) = receiver.as_ref()
-                        {
-                            if method != "chars" {
-                                false
-                            } else if let syn::Expr::Path(syn::ExprPath { path, .. }) =
-                                receiver.as_ref()
-                            {
-                                if let Some(syn::punctuated::Pair::End(path_segment)) =
-                                    path.segments.first()
-                                {
-                                    if let Some(syn::punctuated::Pair::End(syn::FnArg::Captured(
-                                        syn::ArgCaptured {
-                                            pat: syn::Pat::Ident(syn::PatIdent { ident, .. }),
-                                            ..
-                                        },
-                                    ))) = reverse_fn.decl.inputs.first()
-                                    {
-                                        *ident == path_segment.ident
-                                    } else {
-                                        false
-                                    }
-                                } else {
-                                    false
-                                }
-                            } else {
-                                false
-                            }
-                        } else {
-                            false
-                        }
-                    } else {
-                        false
-                    }
-                } else {
-                    false
-                }
-            });
-            if has_standard_solution {
-                output.status = AnalysisStatus::Approve;
-                output
-                    .comments
-                    .push(ReverseStringComment::SuggestDoingBonusTest.to_string())
-            }
-            Ok(())
-        }),
-        Box::new(|ast, output| {
-            let reverse_fn = if let Some(syn::Item::Fn(func)) = ast.items.iter().find(|item| {
-                if let syn::Item::Fn(item_fn) = item {
-                    item_fn.ident == "reverse"
-                } else {
-                    false
-                }
-            }) {
-                func
-            } else {
-                // We can use the unreachable!, since we have already check for
-                // the "reverse" function existence at the analyze function.
-                unreachable!();
-            };
-            // Check if the "reverse" function contains the "input.graphemes(true).rev().collect()"
-            // solution
-            let has_optimal_solution = reverse_fn.block.stmts.iter().any(|stmt| {
-                if let syn::Stmt::Expr(syn::Expr::MethodCall(syn::ExprMethodCall {
-                    method,
-                    receiver,
-                    ..
-                })) = stmt
-                {
-                    if method != "collect" {
-                        false
-                    } else if let syn::Expr::MethodCall(syn::ExprMethodCall {
-                        method,
-                        receiver,
-                        ..
-                    }) = receiver.as_ref()
-                    {
-                        if method != "rev" {
-                            false
-                        } else if let syn::Expr::MethodCall(syn::ExprMethodCall {
-                            method,
-                            receiver,
-                            args,
-                            ..
-                        }) = receiver.as_ref()
-                        {
-                            if method != "graphemes" {
-                                false
-                            } else if let Some(syn::punctuated::Pair::End(syn::Expr::Lit(
-                                syn::ExprLit {
-                                    lit: syn::Lit::Bool(syn::LitBool { value, .. }),
-                                    ..
-                                },
-                            ))) = args.first()
-                            {
-                                *value
-                            } else if let syn::Expr::Path(syn::ExprPath { path, .. }) =
-                                receiver.as_ref()
-                            {
-                                if let Some(syn::punctuated::Pair::End(path_segment)) =
-                                    path.segments.first()
-                                {
-                                    if let Some(syn::punctuated::Pair::End(syn::FnArg::Captured(
-                                        syn::ArgCaptured {
-                                            pat: syn::Pat::Ident(syn::PatIdent { ident, .. }),
-                                            ..
-                                        },
-                                    ))) = reverse_fn.decl.inputs.first()
-                                    {
-                                        *ident == path_segment.ident
-                                    } else {
-                                        false
-                                    }
-                                } else {
-                                    false
-                                }
-                            } else {
-                                false
-                            }
-                        } else {
-                            false
-                        }
-                    } else {
-                        false
-                    }
-                } else {
-                    false
-                }
-            });
-            let has_use_item = if let Some(syn::Item::Use(syn::ItemUse {
-                tree: syn::UseTree::Path(syn::UsePath { ident, tree, .. }),
-                ..
-            })) = ast.items.iter().find(|item| {
-                if let syn::Item::Use(_) = item {
-                    true
-                } else {
-                    false
-                }
-            }) {
-                ident == "unicode_segmentation"
-                    && if let syn::UseTree::Name(syn::UseName { ident }) = tree.as_ref() {
-                        ident == "UnicodeSegmentation"
-                    } else {
-                        false
-                    }
-            } else {
-                false
-            };
-            if has_optimal_solution && has_use_item {
-                output.status = AnalysisStatus::Approve;
-            }
-            Ok(())
-        }),
-    ]
-}
 
 /// Checks if the solution AST contains the necessary function
 /// definition. For `reverse-string` this means checking if the
@@ -304,7 +92,7 @@ impl Analyze for ReverseStringAnalyzer {
             ));
         }
         let mut analysis_output = AnalysisOutput::new(AnalysisStatus::ReferToMentor, vec![]);
-        generate_lints()
+        REVERSE_STRING_LINTS
             .iter()
             .try_for_each(|lint| lint(solution_ast, &mut analysis_output))
             .map(|_| analysis_output)
